ml/requirements.txt
torch==2.2.0
torchvision==0.17.0
timm==0.9.16
numpy==1.26.4
Pillow==10.2.0
scikit-learn==1.4.1.post1
matplotlib==3.8.3
seaborn==0.13.2
pyyaml==6.0.1
tqdm==4.66.2
onnx==1.16.0
onnxruntime==1.17.1
onnx-tf==1.10.0
tensorflow==2.15.0
albumentations==1.4.3
pandas==2.2.1
